+++ timescaledb-cluster/src/main/java/cn/edu/tsinghua/iot/benchmark/timescaledbCluster/TimescaleDB.java
@@ -363,7 +363,7 @@ public class TimescaleDB implements IDatabase {
       recordMap.put(record.getTimestamp(), record.getRecordDataValue());
     }
     sql.append(")");
-    int point = 0;
+    long point = 0;
     int line = 0;
     try (Statement statement = connection.createStatement()) {
       ResultSet resultSet = statement.executeQuery(sql.toString());
